% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{coef.iivw_fit}
\alias{coef.iivw_fit}
\title{Extract Coefficients from iivw_fit}
\usage{
\method{coef}{iivw_fit}(object, ...)
}
\arguments{
\item{object}{An \code{iivw_fit} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Named numeric vector of coefficients.
}
\description{
Extract Coefficients from iivw_fit
}
