% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iivw-package.R
\docType{package}
\name{iivw-package}
\alias{iivw}
\alias{iivw-package}
\title{iivw: Inverse Intensity and Visit Weighting for Longitudinal Data}
\description{
Implements inverse intensity weighting (IIW), inverse probability of treatment weighting (IPTW), and their combination (FIPTIW) for longitudinal studies with irregular visit patterns. IIW corrects for outcome-dependent visit frequency using Andersen-Gill counting-process Cox models (Buzkova & Lumley 2007). IPTW corrects for treatment confounding via propensity scores. FIPTIW combines both corrections (Tompkins, Dubin & Wallace 2025). Provides weight computation and weighted outcome model fitting with GEE or mixed model estimation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tpcopeland/iivw}
  \item Report bugs at \url{https://github.com/tpcopeland/iivw/issues}
}

}
\author{
\strong{Maintainer}: Timothy P Copeland \email{timothy.copeland@ki.se} (\href{https://orcid.org/0000-0003-2801-7928}{ORCID})

}
\keyword{internal}
