% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.iivw_fit}
\alias{print.iivw_fit}
\title{Print iivw_fit Object}
\usage{
\method{print}{iivw_fit}(x, ...)
}
\arguments{
\item{x}{An \code{iivw_fit} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Prints a coefficient table for a fitted IIVW model.
}
