% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.iivw_weights}
\alias{print.iivw_weights}
\title{Print iivw_weights Object}
\usage{
\method{print}{iivw_weights}(x, ...)
}
\arguments{
\item{x}{An \code{iivw_weights} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Prints a summary of the weight diagnostics from \code{\link[=iivw_weight]{iivw_weight()}}.
}
