% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.iivw_fit}
\alias{summary.iivw_fit}
\title{Summarize iivw_fit Object}
\usage{
\method{summary}{iivw_fit}(object, ...)
}
\arguments{
\item{object}{An \code{iivw_fit} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The \code{iivw_fit} object invisibly.
}
\description{
Summarize iivw_fit Object
}
