% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{vcov.iivw_fit}
\alias{vcov.iivw_fit}
\title{Extract Variance-Covariance Matrix from iivw_fit}
\usage{
\method{vcov}{iivw_fit}(object, ...)
}
\arguments{
\item{object}{An \code{iivw_fit} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Variance-covariance matrix.
}
\description{
Extract Variance-Covariance Matrix from iivw_fit
}
